refactor(DTOs): enhance humidity and pressure models, improve air quality response structure
- Update HumidityDetails with additional properties
- Enhance PressureDetails with more detailed measurements
- Refine OpenMeteoAirQualityHourlyResponse structure for better data handling

+++ Weather.Application/DTOs/HumidityDetails.cs
@@ -2,7 +2,7 @@ namespace Weather.Application.DTOs;
 
 public class HumidityDetails
 {
-    // 0 ... 1
-    public double HumidityPercent { get; set; }
+    // 0 ... 100
+    public int HumidityPercent { get; set; }
     public string Summary { get; set; } = null!;
 }
