+++ Weather.Application/DTOs/PressureDetails.cs
@@ -2,6 +2,6 @@ namespace Weather.Application.DTOs;
 
 public class PressureDetails
 {
-    public int PressureHpa { get; set; }
+    public double PressureHpa { get; set; }
     public string Summary { get; set; } = null!;
 }
